Updated with minor fix for
e.   I also create a 3 record CSV file with different Program Andar # (see attached file) and imported to 2017/2018 with and without a replace option.   In either case, no data was found in the search results.

Issue was because the file had empty rows; disregarding rows where the first value is empty from now on